+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/recovery_point/_show.py
@@ -44,6 +44,11 @@ class Show(AAZCommand):
 
         # define Arg Group ""
 
+        _args_schema = cls._args_schema
+        _args_schema.resource_group = AAZResourceGroupNameArg(
+            required=True,
+        )
+
         # define Arg Group "Resource Id Arguments"
 
         _args_schema = cls._args_schema
@@ -60,11 +65,6 @@ class Show(AAZCommand):
             help="Id of the recovery point.",
             required=True,
             id_part="child_name_2",
-        )
-        _args_schema.resource_group = AAZResourceGroupNameArg(
-            arg_group="Resource Id Arguments",
-            help="Name of resource group. You can configure the default group using `az configure --defaults group=<name>`.",
-            required=True,
         )
         _args_schema.vault_name = AAZStrArg(
             options=["--vault-name"],