+++ src/dataprotection/azext_dataprotection/manual/aaz_operations/recovery_point.py
@@ -42,6 +42,10 @@ class RecoveryPointList(_RecoveryPointList):
             import dateutil.parser
             if dateutil.parser.parse(start_time) > dateutil.parser.parse(end_time):
                 logger.warning("The argument --start-time is greater than --end-time.")
+        if start_time and end_time:
+            import dateutil.parser
+            if dateutil.parser.parse(start_time) > dateutil.parser.parse(end_time):
+                logger.warning("The argument --start-time is greater than --end-time.")
 
         rp_filter = ""
         if start_time: