+++ src/dataprotection/azext_dataprotection/manual/custom.py
@@ -820,11 +820,13 @@ def restore_initialize_for_data_recovery(cmd, datasource_type, source_datastore,
 
     manifest = helper.load_manifest(datasource_type)
 
-    # Restore mode (assigned during recovery style earlier) should be supported for the workload
+    # Restore mode (assigned during RP/point-in-time validation earlier) should be supported for the workload
     if manifest is not None and manifest["allowedRestoreModes"] is not None and restore_mode not in manifest["allowedRestoreModes"]:
         raise CLIError(restore_mode + " restore mode is not supported for datasource type " + datasource_type +
                        ". Supported restore modes are " + ','.join(manifest["allowedRestoreModes"]))
 
+    # If the source datastore (type) is allowed for the workload, we start creating the restore request object.
+    # We also check for rehydration priority/duration in here for some reason? It could be shifted out.
     if source_datastore in manifest["policySettings"]["supportedDatastoreTypes"]:
         restore_request["source_data_store_type"] = source_datastore
         if rehydration_priority:
@@ -844,21 +846,14 @@ def restore_initialize_for_data_recovery(cmd, datasource_type, source_datastore,
     datasource_id = None
     # Alternate/Original Location - setting the Target's datasource info accordingly
     if target_resource_id is not None and backup_instance_id is not None:
-        raise CLIError("Please provide either target-resource-id (for alternate location restore) or backup-instance-id \
-                       (for original location restore), not both.")
+        raise CLIError("Please provide either target-resource-id or backup-instance-id not both.")
 
     if target_resource_id is not None:
-        # Verify that the alternate location restore is allowed for datasource type
-        if 'AlternateLocation' not in manifest['allowedRestoreTargetTypes']:
-            raise CLIError('Alternate Location Restore is not allowed for the given DataStoreType \
-                           Please try again with --backup-instance-id parameter.')
+        # No validation for alternate/original location restore, as target_resource_id can be used for both
         datasource_id = target_resource_id
 
     if backup_instance_id is not None:
-        # Verify that the original location restore is allowed for datasource type
-        if 'OriginalLocation' not in manifest['allowedRestoreTargetTypes']:
-            raise CLIError('Original Location Restore is not allowed for the given DataStoreType \
-                           Please try again with --target-resource-id parameter.')
+        # No validation for alternate/original location restore, to be added if understood to be required
         vault_resource_group = helper.get_vault_rg_from_bi_id(backup_instance_id)
         vault_name = helper.get_vault_name_from_bi_id((backup_instance_id))
         backup_instance_name = helper.get_bi_name_from_bi_id(backup_instance_id)
@@ -911,11 +906,14 @@ def restore_initialize_for_data_recovery(cmd, datasource_type, source_datastore,
 
 
 def restore_initialize_for_data_recovery_as_files(target_blob_container_url, target_file_name, datasource_type, source_datastore,
-                                                  restore_location, recovery_point_id=None, point_in_time=None,
+                                                  restore_location, target_resource_id=None,
+                                                  recovery_point_id=None, point_in_time=None,
                                                   rehydration_priority=None, rehydration_duration=15):
 
     restore_request = {}
     restore_mode = None
+
+    # Input Validation and variable-assignment from params for recovery via RP or point-in-time
     if recovery_point_id is not None and point_in_time is not None:
         raise CLIError("Please provide either recovery point id or point in time parameter, not both.")
 
@@ -933,10 +931,14 @@ def restore_initialize_for_data_recovery_as_files(target_blob_container_url, tar
         raise CLIError("Please provide either recovery point id or point in time parameter.")
 
     manifest = helper.load_manifest(datasource_type)
+
+    # Restore mode (assigned during RP/point-in-time validation earlier) should be supported for the workload
     if manifest is not None and manifest["allowedRestoreModes"] is not None and restore_mode not in manifest["allowedRestoreModes"]:
         raise CLIError(restore_mode + " restore mode is not supported for datasource type " + datasource_type +
                        ". Supported restore modes are " + ','.join(manifest["allowedRestoreModes"]))
 
+    # If the source datastore (type) is allowed for the workload, we start creating the restore request object.
+    # We also check for rehydration priority/duration in here for some reason? It could be shifted out.
     if source_datastore in manifest["policySettings"]["supportedDatastoreTypes"]:
         restore_request["source_data_store_type"] = source_datastore
         if rehydration_priority:
@@ -949,6 +951,9 @@ def restore_initialize_for_data_recovery_as_files(target_blob_container_url, tar
         raise CLIError(source_datastore + " datastore type is not supported for datasource type " + datasource_type +
                        ". Supported datastore types are " + ','.join(manifest["policySettings"]["supportedDatastoreTypes"]))
 
+    # Constructing the rest of the restore request object. No further validation is being done.
+    # Currently, restore_target_info.target_details.restore_target_location_type is fixed to AzureBlobs
+    # There is no check currently for ensuring that the manifest's allowedRestoreTargetTypes contains RestoreAsFiles
     restore_request["restore_target_info"] = {}
     restore_request["restore_target_info"]["object_type"] = "RestoreFilesTargetInfo"
     restore_request["restore_target_info"]["restore_location"] = restore_location
@@ -957,6 +962,10 @@ def restore_initialize_for_data_recovery_as_files(target_blob_container_url, tar
     restore_request["restore_target_info"]["target_details"]["url"] = target_blob_container_url
     restore_request["restore_target_info"]["target_details"]["file_prefix"] = target_file_name
     restore_request["restore_target_info"]["target_details"]["restore_target_location_type"] = "AzureBlobs"
+
+    # Mandatory for Cross-subscription restore scenario for OSS
+    if target_resource_id is not None:
+        restore_request["restore_target_info"]["target_details"]["target_resource_arm_id"] = target_resource_id
 
     return restore_request
 
@@ -987,7 +996,7 @@ def restore_initialize_for_item_recovery(cmd, datasource_type, source_datastore,
 
     manifest = helper.load_manifest(datasource_type)
 
-    # Restore mode (assigned during recovery style earlier) should be supported for the workload
+    # Restore mode (assigned during RP/point-in-time validation earlier) should be supported for the workload
     if manifest is not None and manifest["allowedRestoreModes"] is not None and restore_mode not in manifest["allowedRestoreModes"]:
         raise CLIError(restore_mode + " restore mode is not supported for datasource type " + datasource_type +
                        ". Supported restore modes are " + ','.join(manifest["allowedRestoreModes"]))
@@ -996,6 +1005,7 @@ def restore_initialize_for_item_recovery(cmd, datasource_type, source_datastore,
     if manifest is not None and not manifest["itemLevelRecoveyEnabled"]:
         raise CLIError("Specified DatasourceType " + datasource_type + " doesn't support Item Level Recovery")
 
+    # Constructing the rest of the restore request object. No further validation is being done.
     restore_request["source_data_store_type"] = source_datastore
     restore_request["restore_target_info"] = {}
     restore_request["restore_target_info"]["object_type"] = "ItemLevelRestoreTargetInfo"
@@ -1003,9 +1013,18 @@ def restore_initialize_for_item_recovery(cmd, datasource_type, source_datastore,
     restore_request["restore_target_info"]["recovery_option"] = "FailIfExists"
 
     # We set the restore criteria depending on the datasource type and on the prefix pattern/container list as provided
-    # AKS directly uses the restore configuration
+    # AKS directly uses the restore configuration. Currently, the "else" just covers Blobs.
     restore_criteria_list = []
-    if datasource_type != "AzureKubernetesService":
+    if datasource_type == "AzureKubernetesService":
+        if restore_configuration is not None:
+            restore_criteria = restore_configuration
+        else:
+            raise CLIError("Please input parameter restore_configuration for AKS cluster restore.\n\
+                           Use command initialize-restoreconfig for creating the RestoreConfiguration")
+        restore_criteria_list.append(restore_criteria)
+    else:
+        # For non-AKS workloads, we need either a prefix-pattern or a container-list. Accordingly, the restore
+        # criteria's min_matching_value and max_matching_value are set. We need to provide one, but can't provide both
         if container_list is not None and (from_prefix_pattern is not None or to_prefix_pattern is not None):
             raise CLIError("Please specify either container list or prefix pattern.")
 
@@ -1080,34 +1099,20 @@ def restore_initialize_for_item_recovery(cmd, datasource_type, source_datastore,
 
         if container_list is None and from_prefix_pattern is None and to_prefix_pattern is None:
             raise CLIError("Provide ContainersList or Prefixes for Item Level Recovery")
-    else:
-        if restore_configuration is not None:
-            restore_criteria = restore_configuration
-        else:
-            raise CLIError("Please input parameter restore_configuration for AKS cluster restore.\n\
-                           Use command initialize-restoreconfig for creating the RestoreConfiguration")
-        restore_criteria_list.append(restore_criteria)
 
     restore_request["restore_target_info"]["restore_criteria"] = restore_criteria_list
 
     datasource_id = None
     # Alternate/Original Location - setting the Target's datasource info accordingly
     if target_resource_id is not None and backup_instance_id is not None:
-        raise CLIError("Please provide either target-resource-id (for alternate location restore) of backup-instance-id \
-                       (for original location restore), not both.")
+        raise CLIError("Please provide either target-resource-id or backup-instance-id not both.")
 
     if target_resource_id is not None:
-        # Verify that the alternate location restore is allowed for datasource type
-        if 'AlternateLocation' not in manifest['allowedRestoreTargetTypes']:
-            raise CLIError('Alternate Location Restore is not allowed for the given DataStoreType \
-                           Please try again with --backup-instance-id parameter.')
+        # No validation for alternate/original location restore, as target_resource_id can be used for both
         datasource_id = target_resource_id
 
     if backup_instance_id is not None:
-        # Verify that the original location restore is allowed for datasource type
-        if 'OriginalLocation' not in manifest['allowedRestoreTargetTypes']:
-            raise CLIError('Original Location Restore is not allowed for the given DataStoreType \
-                           Please try again with --target-resource-id parameter.')
+        # No validation for alternate/original location restore, to be added if understood to be required
         vault_resource_group = helper.get_vault_rg_from_bi_id(backup_instance_id)
         vault_name = helper.get_vault_name_from_bi_id((backup_instance_id))
         backup_instance_name = helper.get_bi_name_from_bi_id(backup_instance_id)